move ObservedCredResolverCOnfig to a right place
