+++ electron-app/src/renderer/components/credentialsSelection.tsx
@@ -1,5 +1,5 @@
-import { Box, MenuItem, Select, styled, Typography } from '@mui/material'
 import { observer } from 'mobx-react-lite'
+import { Box, MenuItem, Select, styled, Typography } from '@mui/material'
 
 type Option = {
   key: string
