+++ electron-app/src/renderer/store/credResolverStore.ts
@@ -1,11 +1,21 @@
 import _ from 'lodash'
 import { action, computed, flow, makeAutoObservable, makeObservable, observable, toJS } from 'mobx'
 import { singleton } from 'tsyringe'
+import { OBSERVED } from '../../types/mobx'
 import browserLogger from '../logger/browserLogger'
-import { ObservedCredResolverConfig } from '../pages/credResolver/type'
 import { CommonRes, ResultCode } from '../protos/common_pb'
 import { CredResolverConfig, GetCredResolversRes } from '../protos/kubeconfig_service_pb'
 import CredResolverRepository from '../repositories/credResolverRepository'
+
+export type ObservedCredResolverConfig = OBSERVED<
+  CredResolverConfig.AsObject & {
+    resolved?: boolean
+    setResponse?: {
+      resultCode: ResultCode
+      message?: string
+    }
+  }
+>
 
 /**
  * CredResolverStore provides CRUD functionality to application.
@@ -77,6 +87,7 @@ export default class CredResolverStore {
       return
     }
 
+    newConfig.resolved = true
     newConfig.setResponse = {
       resultCode: res.getStatus(),
       message: res.getMessage(),
